Intelligent Recipe Agent Dinner Idea Assistant/source code/code/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
ca-certificates==2024.3.11
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.5
openssl==3.0.13
pillow==10.3.0
pip==24.0
pydantic==2.7.2
pydantic-core==2.18.3
pyqt5==5.15.10
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
python==3.9.19
requests==2.32.3
setuptools==69.5.1
sniffio==1.3.1
sqlite==3.45.3
tqdm==4.66.4
typing-extensions==4.12.0
tzdata==2024a
urllib3==1.25.11
uuid==1.30
vc==14.2
vs2015_runtime==14.29.30133
wheel==0.43.0
